Back/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==23.1.0
browser-cookie3==0.19.1
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
frozenlist==1.4.0
idna==3.4
lz4==4.3.2
multidict==6.0.4
Pillow==9.2.0
pycryptodomex==3.18.0
PyJWT==1.7.1
python-decouple==3.5
python-dotenv==1.0.0
pytz==2022.1
requests==2.31.0
sqlparse==0.4.2
tqdm==4.65.0
tzdata==2023.3
urllib3==2.0.4
yarl==1.9.2
gunicorn==20.1.0
mysqlclient==2.2.0